# speed up grid code paths